Use jellyfish-test-harness to run sync integration tests
Change-type: patch
