+++ package.json
@@ -68,14 +68,12 @@
     "@balena/jellyfish-action-library": "^8.0.22",
     "@balena/jellyfish-core": "^2.0.3",
     "@balena/jellyfish-sync": "^5.0.0",
-    "@balena/jellyfish-queue": "0.0.268",
-    "@balena/jellyfish-worker": "^3.0.0",
+    "@balena/jellyfish-test-harness": "^1.0.0",
     "ava": "^3.15.0",
     "body-parser": "^1.19.0",
     "catch-uncommitted": "^1.6.2",
     "depcheck": "^1.4.0",
     "deplint": "^1.1.3",
-    "errio": "^1.2.2",
     "eslint": "^7.20.0",
     "eslint-config-standard": "^16.0.2",
     "eslint-plugin-import": "^2.22.1",
@@ -85,7 +83,6 @@
     "eslint-plugin-promise": "^4.3.1",
     "express": "^4.17.1",
     "husky": "^5.1.1",
-    "js-combinatorics": "^1.4.5",
     "jsdoc-to-markdown": "^6.0.1",
     "lint-staged": "^10.5.4",
     "morgan": "^1.10.0",
