+++ test/integration/sync/flowdock-translate.spec.js
@@ -5,20 +5,18 @@
  */
 
 const ava = require('ava')
-const scenario = require('./scenario')
 const environment = require('@balena/jellyfish-environment')
+const {
+	syncIntegrationScenario
+} = require('@balena/jellyfish-test-harness')
+const ActionLibrary = require('@balena/jellyfish-action-library')
+const DefaultPlugin = require('../../../lib')
 const TOKEN = environment.integration.flowdock
-const helpers = require('./helpers')
 
-ava.serial.before(async (test) => {
-	await scenario.before(test)
-	await helpers.save(test)
-})
-
-ava.serial.after.always(scenario.after)
-ava.serial.afterEach.always(scenario.afterEach)
-
-scenario.run(ava, {
+syncIntegrationScenario.run(ava, {
+	basePath: __dirname,
+	plugins: [ ActionLibrary, DefaultPlugin ],
+	cards: [ 'thread', 'whisper', 'message' ],
 	integration: require('../../../lib/integrations/flowdock'),
 	scenarios: require('./webhooks/flowdock'),
 	baseUrl: 'https://api.flowdock.com',
